src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.6
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
altair==5.1.1
    # via streamlit
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via langchain
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.6.2
    # via streamlit
cachetools==5.3.1
    # via streamlit
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via streamlit
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.5.14
    # via langchain
exceptiongroup==1.1.3
    # via anyio
fonttools==4.42.1
    # via matplotlib
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.10
    # via gitpython
gitpython==3.1.37
    # via streamlit
google-search-results==2.4.2
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.19.0
    # via altair
jsonschema-specifications==2023.7.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
langchain==0.0.329
    # via
    #   -r requirements.in
    #   langchain-experimental
langchain-experimental==0.0.42
    # via -r requirements.in
langsmith==0.0.63
    # via langchain
lxml==4.9.3
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
marshmallow==3.20.1
    # via dataclasses-json
matplotlib==3.8.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.25.2
    # via
    #   -r requirements.in
    #   altair
    #   contourpy
    #   langchain
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==0.28.0
    # via -r requirements.in
packaging==23.1
    # via
    #   altair
    #   marshmallow
    #   matplotlib
    #   streamlit
pandas==2.1.0
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pillow==9.5.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   streamlit
protobuf==4.24.3
    # via streamlit
pyarrow==14.0.1
    # via streamlit
pydantic==2.3.0
    # via
    #   langchain
    #   langsmith
pydantic-core==2.6.3
    # via pydantic
pydeck==0.8.0
    # via streamlit
pygments==2.16.1
    # via rich
pympler==1.0.1
    # via streamlit
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   streamlit
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0.1
    # via langchain
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   google-search-results
    #   langchain
    #   langsmith
    #   openai
    #   streamlit
rich==13.5.2
    # via streamlit
rpds-py==0.10.3
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.20
    # via langchain
streamlit==1.26.0
    # via -r requirements.in
tabulate==0.9.0
    # via -r requirements.in
tenacity==8.2.3
    # via
    #   langchain
    #   streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.3
    # via streamlit
tqdm==4.66.1
    # via openai
typing-extensions==4.7.1
    # via
    #   altair
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via
    #   pandas
    #   pytz-deprecation-shim
tzlocal==4.3.1
    # via streamlit
urllib3==2.0.7
    # via requests
validators==0.22.0
    # via streamlit
watchdog==3.0.0
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp
zipp==3.16.2
    # via importlib-metadata
